+++ src/main/java/com/itskillerluc/alchemicalbrewery/block/custom/ElementalExtractorBlock.java
@@ -135,10 +135,12 @@ public class ElementalExtractorBlock extends BaseEntityBlock {
 
     @Override
     public void onRemove(@NotNull BlockState pState, @NotNull Level pLevel, @NotNull BlockPos pPos, @NotNull BlockState pNewState, boolean pIsMoving) {
-        BlockEntity blockEntity = pLevel.getBlockEntity(pPos);
-        super.onRemove(pState, pLevel, pPos, pNewState, pIsMoving);
-        if (blockEntity instanceof ElementalExtractorTile extractorTile) {
-            extractorTile.drops();
+        if (pState.getBlock() != pNewState.getBlock()) {
+            BlockEntity blockEntity = pLevel.getBlockEntity(pPos);
+            if (blockEntity instanceof ElementalExtractorTile extractorTile) {
+                extractorTile.drops();
+            }
+            super.onRemove(pState, pLevel, pPos, pNewState, pIsMoving);
         }
     }
